refactor(net): extract shared netFD and address types into NetCommon
Pure move of the socket-address types, sockaddr conversions, and the
internal FD (netFD) owner out of TCP into a shared NetCommon module,
mirroring Go's net package where one netFD backs every socket type.
TCP re-exposes the public names unchanged via imports. _new_netfd now
derives is_stream/zero_read_is_eof from the socket type, and
open_net_fd! generalizes open_tcp_fd! (which delegates). No behavior
change; TCP tests pass unchanged.
