+++ src/Reseau.jl
@@ -12,6 +12,7 @@ export TCP, TLS
 include("0_compat.jl")
 include("1_socket_ops.jl")
 include("2_iopoll.jl")
+include("3_netcommon.jl")
 include("3_tcp.jl")
 include("4_host_resolvers.jl")
 include("5_socks.jl")
